+++ app/issues/[id]/EditIssueBtn.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { Button } from '@radix-ui/themes';
 import { Pencil2Icon } from '@radix-ui/react-icons';
 
-const EditIssueBtn = ({issueId}:{issueId:number}) => {
+const EditIssueBtn = ({ issueId }: { issueId: number }) => {
     return (
         <Button>
             <Pencil2Icon />