+++ app/issues/[id]/IssueDetails.tsx
@@ -11,7 +11,7 @@ const IssueDetails = ({ issue }: { issue: Issue }) => {
                 <IssueStatusBadge status={issue.status} />
                 <Text><b>Created on: </b> {issue.createdAt.toDateString()}</Text>
             </Flex>
-            <Card className='prose mt-5 !w-full'>
+            <Card className='prose max-w-full' mt={'4'}>
                 <Markdown>{issue.description}</Markdown>
             </Card>
         </>