+++ app/issues/[id]/page.tsx
@@ -1,6 +1,7 @@
 import prisma from '@/prisma/client';
-import { Box, Grid } from '@radix-ui/themes';
+import { Box, Flex, Grid } from '@radix-ui/themes';
 import { notFound } from 'next/navigation';
+import DeleteIssueBtn from './DeleteIssueBtn';
 import EditIssueBtn from './EditIssueBtn';
 import IssueDetails from './IssueDetails';
 interface Props {
@@ -18,12 +19,15 @@ const IssueDetailPage = async ({ params }: Props) => {
     }
 
     return (
-        <Grid columns={{ initial: '1', md: '2' }} gap={'5'}>
-            <Box>
+        <Grid columns={{ initial: '1', sm: '5' }} gap={'5'}>
+            <Box className='md:col-span-4'>
                 <IssueDetails issue={issueDetails} />
             </Box>
             <Box>
-                <EditIssueBtn issueId={issueDetails.id} />
+                <Flex direction={'column'} gap={'4'}>
+                    <EditIssueBtn issueId={issueDetails.id} />
+                    <DeleteIssueBtn issueId={issueDetails.id} />
+                </Flex>
             </Box>
         </Grid>
     )