+++ documentation/changelog.md
@@ -1,13 +1,18 @@
 ### Tracker.io - A Issue Tracker app made with Nextjs.
 
 
-> 1.3.4 - Bug Fixes : 
+> 1.3.5 - Bug Fixes 4 & Improvements:
+- Fixed the layout using radixui container.
+- Added DeleteIssueBtn.
+- Made the edit page a bit lil more responsive.
+
+> 1.3.4 - Bug Fixes 3: 
 - Fixed navigator not found errors due to static imports in server page. so added dynamic imports and turned off ssr to avoid making that a static page.
 - Fixed loading animations & skeletons.
 - Refactored the loading skeleton for issueform so that it can be reused.
 - Removed dynamic import from simpleMDE editor and added it to whole page instead as mde editor was loading lazily and other elements like title is loading eagerly thereby distrubing the loading experience..
 
-> 1.3.3 - Bug Fixes : Fixed static rendering issue of issues page.
+> 1.3.3 - Bug Fixes 2: Fixed static rendering issue of issues page.
 - This bug led to not updating the issues after adding/updating the issues.
 - so kind of added a manual page refresh.
 - Fixed api schema name import and its  usage.